pyproject.toml
[tool.poetry]
name = "ride-planner"
version = "0.1.0"
description = "Simple ai ride planner"
authors = ["winner-team"]


[tool.poetry.dependencies]  
python = ">=3.10.0,<3.13"
crewai = "0.1.24"
unstructured = '==0.10.25'
pyowm = '3.3.0'
tools = "^0.1.9"
langchain = "0.0.354"
langchain-groq = "0.0.1"
python-dotenv = "1.0.0"
python-decouple = "^3.8"
config = "^0.5.1"
decouple = "^0.0.7"



[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"